WIP: correcting seat determining math
the mathematics of the connectivity matrix and its use to calculate the
energy (penalty) of the system are corrected

to determine the correctness of the energy in the case of the
connectivity matrix and interaction matrix, the connectivity matrix for
a circular arrangement (circulant off-diagonals matrix) is evaluated and
found to the be same as calculated by pairwise linear enumeration. it
remains to implement an efficient routine for updating the intraction
matrix for only those nearest neighbors affected by a swap rather than
reevaluating the entire matrix. some checks are also made on the
connectivity matrix (the points are correctly identified as
corner/edge/interior)