WIP: calculation of seat places by correlation
optimum found is still not correct by intuition, sometimes having empty
seats on edges or corners. tests for the connectivity matrix and its
logic construction indicate it is correct. otherwise problems with
temperature or number of iterations. there must be both theoretical and
empirical calculations to be done to determine optimum or rule of thumb
temperature, convergence criteria, and so on