first pass implementation of by-category projection
the routine is inefficient. it just boolean masks the array for each
date. a simple improvement would be to calculate the cumulative sum once
for each category and then select the rows to add together depending on
date. the cumulative sum on a concatenated array can't be used when the
predicted number of total points in each category is based on to-date
fractional performance, so the routine has to be different than the
projection based on total poins

function refactoring is in progress. across all files the function definitions
are not consistent in what they input and output, often using global
variables and mixing calculation scripts with visualization scripts