fix for true grades and made consistent funcs
the weighted grades matrix, giving the fraction, does not give the true
grade matrix, which is the sum normalized by the total. so a true grades
matrix was made. also the term matrix is not needed, since it isn't
being used for linear algebra or matrix operations, and so is dropped.
functions consistently take the same inputs and generally return either
pandas or matplotlib objects, and only operate on them, rather than
doing other things like saving files