updated make-seat-assign with efficient calc
in the general case the minimum isn't known, an evaluation is made of
the history of the energy changes. this is done relative to the provided
normalization constant (called kbt as in thermodynamics for correctly
dimensioned temperature). otherwise it quites after the fourth power of
the number of elements, which was chosen without reason