added check by random sampling
microstates (that is, all of the positions) are randomized to check if a
solution. note however that randomly sampling some number of times never
guarantees finding the optimum, and it doesn't give high confidence
unless you do it enough times that you effectively sample the space,
which removes the whole point of making a search routine to find the
optimum. hence this is only a sanity check.